chore(deps): bump versions
