+++ package.json
@@ -20,12 +20,12 @@
   "license": "UNLICENSED",
   "bugs": "https://github.com/fintruth/fintruth-sdk/issues",
   "devDependencies": {
-    "@types/jest": "^24.0.6",
+    "@types/jest": "^24.0.9",
     "@types/node": "^11.9.5",
     "@typescript-eslint/eslint-plugin": "^1.4.2",
     "@typescript-eslint/parser": "^1.4.2",
     "eslint": "^5.14.1",
-    "eslint-config-prettier": "^4.0.0",
+    "eslint-config-prettier": "^4.1.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-react": "^7.0.2",
     "eslint-import-resolver-typescript": "^1.1.1",
@@ -45,7 +45,7 @@
     "jest": "^24.1.0",
     "jest-circus": "^24.1.0",
     "jest-config": "^24.1.0",
-    "lerna": "^3.13.0",
+    "lerna": "^3.13.1",
     "lint-staged": "^8.1.4",
     "npm-run-all": "^4.1.5",
     "prettier": "^1.16.4",
