+++ packages/client/package.json
@@ -39,11 +39,10 @@
     "@reach/dialog": "^0.1.3",
     "@reach/menu-button": "^0.1.8",
     "@reach/router": "^1.2.1",
-    "apollo-cache-inmemory": "^1.4.3",
-    "apollo-client": "^2.4.13",
+    "apollo-cache-inmemory": "^1.5.0",
+    "apollo-client": "^2.5.0",
     "apollo-link": "^1.2.8",
     "apollo-link-error": "^1.1.7",
-    "apollo-link-state": "^0.4.2",
     "apollo-upload-client": "^10.0.0",
     "body-parser": "^1.18.3",
     "compression": "^1.7.3",
@@ -57,7 +56,7 @@
     "prop-types": "^15.7.2",
     "ramda": "^0.26.1",
     "react": "^16.8.3",
-    "react-apollo": "^2.4.1",
+    "react-apollo": "^2.5.0",
     "react-dom": "^16.8.3",
     "reactabular-table": "^8.14.0",
     "serialize-javascript": "^1.6.1",
@@ -98,9 +97,9 @@
     "@types/loadable__component": "^5.6.0",
     "@types/loadable__server": "^5.2.1",
     "@types/node-fetch": "^2.1.6",
-    "@types/ramda": "^0.25.50",
+    "@types/ramda": "^0.25.51",
     "@types/reach__router": "^1.2.3",
-    "@types/react": "^16.8.4",
+    "@types/react": "^16.8.5",
     "@types/react-dom": "^16.8.2",
     "@types/serialize-javascript": "^1.5.0",
     "@types/storybook__addon-a11y": "^3.3.1",
@@ -111,7 +110,7 @@
     "@types/storybook__addon-viewport": "^4.1.0",
     "@types/storybook__react": "^4.0.1",
     "@types/styled-components": "4.1.10",
-    "@types/yup": "^0.26.9",
+    "@types/yup": "^0.26.10",
     "apollo-link-logger": "^1.2.3",
     "babel-jest": "^24.1.0",
     "babel-loader": "^8.0.5",
