+++ packages/server/package.json
@@ -25,7 +25,7 @@
   "bugs": "https://github.com/fintruth/fintruth-sdk/issues",
   "dependencies": {
     "@fintruth-sdk/shared": "^0.1.0",
-    "apollo-server-express": "^2.4.6",
+    "apollo-server-express": "^2.4.8",
     "base64url": "^3.0.1",
     "bcrypt": "^3.0.4",
     "compression": "^1.7.3",
@@ -63,7 +63,7 @@
     "@types/jsonwebtoken": "8.3.0",
     "@types/morgan": "^1.7.35",
     "@types/qrcode": "^1.3.1",
-    "@types/ramda": "^0.25.50",
+    "@types/ramda": "^0.25.51",
     "@types/speakeasy": "^2.0.4",
     "@types/split": "^1.0.0",
     "@types/uuid": "^3.4.4",
