+++ packages/validation/package.json
@@ -31,7 +31,7 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@types/yup": "^0.26.9",
+    "@types/yup": "^0.26.10",
     "@types/zxcvbn": "^4.4.0"
   },
   "engines": {
